Remove autoload and sign caching
Per critique, there is no point in autoloading anything in this plugin,
as it is intended to always be available. The sign caching approach
has also been canned, as without a proper "sign change" autocmd event,
there is no foolproof way to ensure the sign cache is always up to date.