Add experimental helper function for name maps
Experimental user-interface to creating mappings that only jump to
a signs with a name that matches a given list of patterns. Given a key
sequence to map, an action, and a list of sign name patterns, generate
a normal and visual mode mapping for the jump motion specified by
action.